+++ src/main/java/de/uni/leipzig/IR15/Importer/Neo4JImporter.java
@@ -80,7 +80,7 @@ public class Neo4JImporter extends Importer {
 	      while (rs.next()) {
 	        Integer w1_id 	= rs.getInt("w1_id");
 	        Integer w2_id 	= rs.getInt("w2_id");
-	        Integer sig 	= rs.getInt("sig");
+	        Double sig 		= rs.getDouble("sig");
 	        Integer freq 	= rs.getInt("freq");
 	        	        
 	        Node source = nodeIndex.get("w_id", w1_id).getSingle();
